refactor: extract tilt projection helpers to dedicated module
- Created runtime/projections/tilt.py and migrated logic from runtime/tilt_projection.py.
- Removed _project_tilts_to_tangent_array and _project_tilts_axisymmetric_about_center static methods from Minimizer.
- Updated Minimizer and tests to use module-level projection functions.
- Verified behavior with existing projection and minimizer tests.
